requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.4
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-environ==0.10.0
django-js-asset==2.1.0
django-resized==1.0.2
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2==2.9.7
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
